cleanup code add tests stage to pipeline
