+++ package.json
@@ -17,7 +17,7 @@
     "pretest": "npm run lint && npm run validate-templates",
     "lint": "eslint bin lib test index.js",
     "test": "docker build -q -t ecs-watchbot -f test/Dockerfile ./ && docker run -t ecs-watchbot npm run test-container",
-    "test-container": "tape test/*.test.js && jest test/*.spec.js",
+    "test-container": "tape test/*.test.js",
     "validate-templates": "tape test/template.validation.js",
     "update-jest-snapshots": "jest -u test/*.spec.js",
     "coverage": "nyc --reporter html tape test/*.test.js && opener coverage/index.html",
